cli: Use types instead of instance names, and update for new Image instances
Turns out when Webpack optimises the code it changes all the class names to be
as short as possible, which means all the string matches against class names
fail!  So we have to ditch that and identify instances by other means.  This is
probably better anyway, because it will allow us to send the same Image
instance to either an image editor or a palette editor as required.